modif:修改接口，支持浮点数的计算，但是现在有个问题是浮点数的乘除以及负数的乘除都不能100%的通过测试

+++ Weather_App/MMAST.h
@@ -8,6 +8,7 @@
 
 #import <Foundation/Foundation.h>
 #import "MMObjDescribeAble.h"
+#import <UIKit/UIKit.h>
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -15,7 +16,7 @@ NS_ASSUME_NONNULL_BEGIN
 @interface MMAST : NSObject<MMObjDescribeAble>
 
 // subclass override
-- (NSInteger) visit;
+- (CGFloat) visit;
 - (NSString *) toLispStyle;
 - (NSString *) toBackStyle;
 @end
